Restructure internal modules
- Move `compiler` module to root
- Move prover and verifier modules under compiler
- Move compress module under composer
- Move constraint_system module under composer
- Move permutation module under composer
