+++ src/compiler/verifier.rs
