+++ src/composer/compress/hades.rs
