+++ src/composer/permutation/constants.rs
